Refactor tests to run in parallel
This commit restructures the existing tests to ensure there's no overlap
between resources required for each test. That process involves ensuring
all dependent network resources are generated with random IDs and that
the fixtures directory is the main entry point for tests. This is being
done to setup the address module for a CI pipeline which will
parallelize test execution.